Remove obsolete %t mode line construct documentation
src/buffer.c (mode-line-format): Remove %t from docstring.

This documentation had been previously removed in 1999 in
commit 08de62001945d4dd32a3d6af44da05804cb296d5 but was
subsequently restored again in a different form in 2006 in
commit dafbe726b89a36010f3f24bb1efe4ace0a5f7f0f as the
construct was still supported on some systems.  However it
looks like it hasn't done what the documentation indicated
since commit 05c652517317d10690aaf0a6aa0bc876382b9d82 in 2011,
and the final remnant of this functionality was removed in
commit 8549f9e89bd9288c4c709d183a5bf8f07dbeed3d in 2013; so
it seems clear that it's no longer supported and shouldn't
be documented now.