Add basic project-build-dir implementation
project.el: Add a new custom project-extra-info and a generic with same
name in order to specialize some commands when possible.

Fix project-recompile

I am not 100% happy of this change because the compile-command value is
not stored as project-local, but as buffer-local.

Properly fix project-local parameters.

This also solves the interactive compilation in the previous commit